+++ src/wasm/body.rs
@@ -19,8 +19,12 @@ pub struct Body {
 
 enum Inner {
     Bytes(Bytes),
+    /// MultipartForm holds a multipart/form-data body.
     #[cfg(feature = "multipart")]
-    Multipart(Form),
+    MultipartForm(Form),
+    /// MultipartPart holds the body of a multipart/form-data part.
+    #[cfg(feature = "multipart")]
+    MultipartPart(Bytes),
 }
 
 impl Body {
@@ -32,7 +36,9 @@ impl Body {
         match &self.inner {
             Inner::Bytes(bytes) => Some(bytes.as_ref()),
             #[cfg(feature = "multipart")]
-            Inner::Multipart(_) => None,
+            Inner::MultipartForm(_) => None,
+            #[cfg(feature = "multipart")]
+            Inner::MultipartPart(bytes) => Some(bytes.as_ref()),
         }
     }
     pub(crate) fn to_js_value(&self) -> crate::Result<JsValue> {
@@ -44,9 +50,18 @@ impl Body {
                 Ok(js_value.to_owned())
             }
             #[cfg(feature = "multipart")]
-            Inner::Multipart(form) => {
+            Inner::MultipartForm(form) => {
                 let form_data = form.to_form_data()?;
                 let js_value: &JsValue = form_data.as_ref();
+                Ok(js_value.to_owned())
+            }
+            #[cfg(feature = "multipart")]
+            Inner::MultipartPart(body_bytes) => {
+                let body_bytes: &[u8] = body_bytes.as_ref();
+                let body_uint8_array: Uint8Array = body_bytes.into();
+                let body_array = js_sys::Array::new();
+                body_array.push(&body_uint8_array);
+                let js_value: &JsValue = body_array.as_ref();
                 Ok(js_value.to_owned())
             }
         }
@@ -56,7 +71,23 @@ impl Body {
     #[cfg(feature = "multipart")]
     pub(crate) fn from_form(f: Form) -> Body {
         Self {
-            inner: Inner::Multipart(f),
+            inner: Inner::MultipartForm(f),
+        }
+    }
+
+    /// into_part turns a regular body into the body of a mutlipart/form-data part.
+    #[cfg(feature = "multipart")]
+    pub(crate) fn into_part(self) -> Body {
+        match self.inner {
+            Inner::Bytes(bytes) => Self {
+                inner: Inner::MultipartPart(bytes),
+            },
+            Inner::MultipartForm(form) => Self {
+                inner: Inner::MultipartForm(form),
+            },
+            Inner::MultipartPart(bytes) => Self {
+                inner: Inner::MultipartPart(bytes),
+            },
         }
     }
 
@@ -64,7 +95,9 @@ impl Body {
         match &self.inner {
             Inner::Bytes(bytes) => bytes.is_empty(),
             #[cfg(feature = "multipart")]
-            Inner::Multipart(form) => form.is_empty(),
+            Inner::MultipartForm(form) => form.is_empty(),
+            #[cfg(feature = "multipart")]
+            Inner::MultipartPart(bytes) => bytes.is_empty(),
         }
     }
 
@@ -74,7 +107,11 @@ impl Body {
                 inner: Inner::Bytes(bytes.clone()),
             }),
             #[cfg(feature = "multipart")]
-            Inner::Multipart(_) => None,
+            Inner::MultipartForm(_) => None,
+            #[cfg(feature = "multipart")]
+            Inner::MultipartPart(bytes) => Some(Self {
+                inner: Inner::MultipartPart(bytes.clone()),
+            }),
         }
     }
 }
@@ -130,7 +167,8 @@ impl fmt::Debug for Body {
 
 #[cfg(test)]
 mod tests {
-    // use js_sys::{Array, Uint8Array};
+    use crate::Body;
+    use js_sys::Uint8Array;
     use wasm_bindgen::prelude::*;
     use wasm_bindgen_test::*;
 
@@ -146,16 +184,12 @@ mod tests {
 
     #[wasm_bindgen_test]
     async fn test_body() {
-        use crate::Body;
-
         let body = Body::from("TEST");
         assert_eq!([84, 69, 83, 84], body.as_bytes().unwrap());
     }
 
     #[wasm_bindgen_test]
-    async fn test_body_js() {
-        use crate::Body;
-
+    async fn test_body_js_static_str() {
         let body_value = "TEST";
         let body = Body::from(body_value);
 
@@ -175,5 +209,83 @@ mod tests {
             .expect("could not get request body as text");
 
         assert_eq!(text.as_string().expect("text is not a string"), body_value);
+    }
+    #[wasm_bindgen_test]
+    async fn test_body_js_string() {
+        let body_value = "TEST".to_string();
+        let body = Body::from(body_value.clone());
+
+        let mut init = web_sys::RequestInit::new();
+        init.method("POST");
+        init.body(Some(
+            body.to_js_value()
+                .expect("could not convert body to JsValue")
+                .as_ref(),
+        ));
+
+        let js_req = web_sys::Request::new_with_str_and_init("", &init)
+            .expect("could not create JS request");
+        let text_promise = js_req.text().expect("could not get text promise");
+        let text = crate::wasm::promise::<JsValue>(text_promise)
+            .await
+            .expect("could not get request body as text");
+
+        assert_eq!(text.as_string().expect("text is not a string"), body_value);
+    }
+
+    #[wasm_bindgen_test]
+    async fn test_body_js_static_u8_slice() {
+        let body_value: &'static [u8] = b"\x00\x42";
+        let body = Body::from(body_value);
+
+        let mut init = web_sys::RequestInit::new();
+        init.method("POST");
+        init.body(Some(
+            body.to_js_value()
+                .expect("could not convert body to JsValue")
+                .as_ref(),
+        ));
+
+        let js_req = web_sys::Request::new_with_str_and_init("", &init)
+            .expect("could not create JS request");
+
+        let array_buffer_promise = js_req
+            .array_buffer()
+            .expect("could not get array_buffer promise");
+        let array_buffer = crate::wasm::promise::<JsValue>(array_buffer_promise)
+            .await
+            .expect("could not get request body as array buffer");
+
+        let v = Uint8Array::new(&array_buffer).to_vec();
+
+        assert_eq!(v, body_value);
+    }
+
+    #[wasm_bindgen_test]
+    async fn test_body_js_vec_u8() {
+        let body_value = vec![0u8, 42];
+        let body = Body::from(body_value.clone());
+
+        let mut init = web_sys::RequestInit::new();
+        init.method("POST");
+        init.body(Some(
+            body.to_js_value()
+                .expect("could not convert body to JsValue")
+                .as_ref(),
+        ));
+
+        let js_req = web_sys::Request::new_with_str_and_init("", &init)
+            .expect("could not create JS request");
+
+        let array_buffer_promise = js_req
+            .array_buffer()
+            .expect("could not get array_buffer promise");
+        let array_buffer = crate::wasm::promise::<JsValue>(array_buffer_promise)
+            .await
+            .expect("could not get request body as array buffer");
+
+        let v = Uint8Array::new(&array_buffer).to_vec();
+
+        assert_eq!(v, body_value);
     }
 }